api: move host scanning to bus
